+++ apps/server/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @objectstack/example-host
 
+## 4.0.4
+
+### Patch Changes
+
+- Updated dependencies [326b66b]
+  - @objectstack/spec@4.0.4
+  - @example/app-crm@4.0.4
+  - @example/app-todo@4.0.4
+  - @example/plugin-bi@4.0.4
+  - @objectstack/metadata@4.0.4
+  - @objectstack/objectql@4.0.4
+  - @objectstack/driver-memory@4.0.4
+  - @objectstack/driver-turso@4.0.4
+  - @objectstack/plugin-audit@4.0.4
+  - @objectstack/plugin-auth@4.0.4
+  - @objectstack/plugin-hono-server@4.0.4
+  - @objectstack/plugin-security@4.0.4
+  - @objectstack/plugin-setup@4.0.4
+  - @objectstack/runtime@4.0.4
+  - @objectstack/service-ai@4.0.4
+  - @objectstack/service-analytics@4.0.4
+  - @objectstack/service-automation@4.0.4
+  - @objectstack/service-feed@4.0.4
+  - @objectstack/hono@4.0.4
+
 ## 4.0.3
 
 ### Patch Changes