+++ examples/app-crm/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @objectstack/example-crm
 
+## 4.0.4
+
+### Patch Changes
+
+- Updated dependencies [326b66b]
+  - @objectstack/spec@4.0.4
+
 ## 4.0.3
 
 ### Patch Changes