+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @objectstack/cli
 
+## 4.0.4
+
+### Patch Changes
+
+- Updated dependencies [326b66b]
+  - @objectstack/spec@4.0.4
+  - @objectstack/client@4.0.4
+  - @objectstack/core@4.0.4
+  - @objectstack/objectql@4.0.4
+  - @objectstack/driver-memory@4.0.4
+  - @objectstack/plugin-hono-server@4.0.4
+  - @objectstack/plugin-setup@4.0.4
+  - @objectstack/rest@4.0.4
+  - @objectstack/runtime@4.0.4
+  - @objectstack/service-ai@4.0.4
+
 ## 4.0.3
 
 ### Patch Changes