+++ packages/objectql/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @objectstack/objectql
 
+## 4.0.4
+
+### Patch Changes
+
+- Updated dependencies [326b66b]
+  - @objectstack/spec@4.0.4
+  - @objectstack/core@4.0.4
+  - @objectstack/types@4.0.4
+
 ## 4.0.3
 
 ### Patch Changes