+++ packages/services/service-cache/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @objectstack/service-cache
 
+## 4.0.4
+
+### Patch Changes
+
+- Updated dependencies [326b66b]
+  - @objectstack/spec@4.0.4
+  - @objectstack/core@4.0.4
+
 ## 4.0.3
 
 ### Patch Changes