+++ packages/vscode-objectstack/CHANGELOG.md
@@ -1,5 +1,7 @@
 # objectstack-vscode
 
+## 4.0.4
+
 ## 4.0.3
 
 ## 4.0.2